api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
openai>=1.0.0
pydantic>=1.8.0
slowapi>=0.1.4
redis>=4.0.0
python-multipart>=0.0.5
llama-index>=0.9.0
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-qdrant>=0.1.0
llama-index-readers-file>=0.1.0
qdrant-client>=1.7.0
pypdf>=3.17.0
python-magic>=0.4.27
pillow>=10.0.0
numpy>=1.24.0